Commit 19 - Fixed Layout issue!
Finally. It looks good in all the viewport sizes. (320px-1024px). Just
need to complete the functionality of the local storage, the cancel
button at the bottom, fix the notification bell at the top to be in the
correct position at tablet and desktop view while also having
functionality and this project should be finished!